% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILnc.R
\name{ILnc}
\alias{ILnc}
\title{ILnc}
\usage{
ILnc(x, y, perm = 0)
}
\arguments{
\item{x}{A matrix with samples as colnames,Genes (Ensembel ID) as the rownames}

\item{y}{1 or more names of the 14 cell types:("B","DC","Granulocytes","M0","M1","M2","Macrophage","Monocyte","Neutrophil","NK","CD8 T","CD4 T","CD3 T","Treg")}

\item{perm}{Number of permutations}
}
\value{
A matrix of immune cells prediction result
}
\description{
This package is designed to predict the infiltration of 14 immune cells in Mixed organization.
}
\examples{
y=c("B","M0","M1")
}
